+++ src/app/pages/plan-estudios/creacion-plan-estudios/creacion-plan-estudios.component.scss
@@ -29,8 +29,8 @@ fieldset {
 
 
 .alinear-center-espacio-doble {
-    padding-left: 15%;
-    padding-right: 15%;
+    padding-left: 25%;
+    padding-right: 25%;
     padding-top: 1rem;
     padding-bottom: 1rem;
     display: flex;
