+++ src/app/pages/plan-estudios/edicion-plan-estudios/edicion-plan-estudios.component.html
@@ -3,7 +3,7 @@ nbSpinnerMessage="{{ 'GLOBAL.cargando' | translate }}" [ngSwitch]="vista">
     <nb-card-header>
         <div class="row" align="center">
             <div class="col">
-                {{ 'plan_estudios.crear' | translate }}
+                {{ 'plan_estudios.editar' | translate }}
             </div>
         </div>
     </nb-card-header>
@@ -17,7 +17,7 @@ nbSpinnerMessage="{{ 'GLOBAL.cargando' | translate }}" [ngSwitch]="vista">
                 <ng2-smart-table [settings]="tbPlanesEstudio" [source]="dataPlanesEstudio">
                 </ng2-smart-table>
                 <div class="alinear-center espacio-doble">
-                    <button mat-raised-button color="primary" (click)="nuevoPlanEstudio()">{{ 'plan_estudios.crear' | translate }}</button>
+                    <button mat-raised-button color="primary" (click)="organizarPlanesEstudio()">{{ 'plan_estudios.organizar' | translate }}</button>
                     <button mat-raised-button color="primary" (click)="organizarPlanesEstudio()" [hidden]="habilitadoGenerarPlan">{{ 'plan_estudios.generar_plan' | translate }}</button>
                 </div>
             </fieldset>
@@ -33,7 +33,6 @@ nbSpinnerMessage="{{ 'GLOBAL.cargando' | translate }}" [ngSwitch]="vista">
                         <ng-template matStepLabel>{{ 'espacios_academicos.select_proyecto' | translate }}</ng-template>
                         <dinamic-form-group [defineForm]="formPlanEstudio" (subsChanges)="cambioEn($event)" (updatedForm)="actualizarForm($event)"></dinamic-form-group>
                         <div class="alinear-center">
-                            <button mat-raised-button (click)="limpiar()">{{ 'GLOBAL.limpiar' | translate }}</button>
                             <button mat-raised-button [color]="formGroupPlanEstudio.valid ? 'primary' : ''" (click)="guardar(stepper)">{{ 'GLOBAL.guardar_continuar' | translate }}</button>
                         </div>
                     </mat-step>
